refactor(cli): split loader layer out of main.rs
Second incremental step toward decomposing the original 6.3k-line
main.rs. After the render-layer split (PR #70), this moves the
adapter-loading layer into a new loader.rs:

- StateValuationsMap type alias
- parse_context_file (CTXDSL file -> ContextDoc parse helper)
- log_adapter_output, log_adapter_output_with_dir (private debug
  output for adapter intermediate states)
- load_with_adapter_mode (primary adapter dispatch)
- load_context_documents, load_context_documents_mode
- realize_documents (wraps mununu_core::realize_context)
- print_context_structure, print_ctxdsl_output
- PreparedEvalContext, EvalContextParams<'a>, prepare_eval_context
  (the eval-flow preamble extracted in PR #68; sits naturally with
  the loader since it consumes load_context_documents_mode +
  print_ctxdsl_output and produces a realized context)

main.rs: 4,940 -> 4,557 lines (-383). loader.rs: 410 lines. Items
used from main.rs are pub(crate); log_adapter_output and its
_with_dir variant stay private to loader.rs (only callers are
inside load_with_adapter_mode).

Deliberately left in main.rs:
- resolve_formula_name (semantically part of the eval-handler
  preamble; will move with the eval handler in the next session)
- build_context_summary (entangled with handler-level
  AutomatonSummaryOutput / ContextSummaryOutput types; moves with
  the context handler)

prepare_eval_context calls crate::resolve_formula_name across the
module boundary, which is valid (same crate) and not cyclic.

All 1,611 tests pass; clippy -D warnings clean.

This PR depends on PR #70 (render-layer split) and should merge after
it. Targets main directly; will auto-rebase to a clean loader-only
diff once #70 lands.
